fix: IC-136 推 CI 前的两处类型收口（A 与 C 的测试各一处）
推送前静态复核抓到两处会编译失败或可能失败的写法，各归其子项，未改任何断言语义：

- A：`Mirror.Child` 是元组（`(label: String?, value: Any)`），Swift 不支持指向
  元组成员的 key path，`compactMap(\.label)` 编译不过；改用闭包
  `compactMap { $0.label }`。
- C：`S2PendingDeletionMark.circleOpacity` 是 `Double`，采样值是 `CGFloat`，
  `XCTAssertEqual(_:_:accuracy:)` 三个参数须同类型；显式 `CGFloat(...)` 包一层，
  不依赖 CGFloat ↔ Double 的隐式转换在泛型推导位置成立。

不 amend 前三个提交（本卡范围外），故单列此提交。

+++ PhotoCleanupMVETests/S2ActionBarWiringTests.swift
@@ -656,7 +656,7 @@ final class S2ActionBarWiringTests: XCTestCase {
         XCTAssertEqual(background.white, 0, accuracy: 0.001, "底应为黑")
         XCTAssertEqual(
             background.alpha,
-            S2PendingDeletionMark.circleOpacity,
+            CGFloat(S2PendingDeletionMark.circleOpacity),
             accuracy: 0.001
         )
         let foreground = grayscale(S2CenterIndicatorView.foregroundColor, .light)
